+++ cnccode/controller_classes/CTActivity.inc.php
@@ -1167,7 +1167,8 @@ class CTActivity extends CTCNC
         $customerID,
         $contractCustomerItemID,
         $templateName = 'ActivityCreate6',
-        $blockName = 'contractBlock'
+        $blockName = 'contractBlock',
+        bool $linkedToSalesOrder = false
     )
     {
         $buCustomerItem = new BUCustomerItem($this);
@@ -1232,7 +1233,8 @@ class CTActivity extends CTCNC
                     'optGroupClose'          => $optGroupClose,
                     'optGroupCloseLast'      => $dsContract->rowCount() == $currentRow ? '</optgroup>' : 'null',
                     'prepayContract'         => $dsContract->getValue(DBEJContract::itemTypeID) == 57,
-                    'isDisabled'             => $dsContract->getValue(DBEJContract::allowSRLog) ? null : 'disabled'
+                    'isDisabled'             => $dsContract->getValue(DBEJContract::allowSRLog) ? null : 'disabled',
+                    'disabled'               => $linkedToSalesOrder ? 'disabled' : null
                 )
             );
             $this->template->parse(
@@ -6152,11 +6154,12 @@ class CTActivity extends CTCNC
                 ) < $dsHeader->getValue(DBEHeader::srPromptContractThresholdHours))
 
         ) {
-
             $this->contractDropdown(
                 $dsCallActivity->getValue(DBEJCallActivity::customerID),
                 $this->getParam('contractCustomerItemID'),
-                'ServiceRequestFixedEditContractDropdown'
+                'ServiceRequestFixedEditContractDropdown',
+                'contractBlock',
+                !!$dsCallActivity->getValue(DBEJCallActivity::linkedSalesOrderID)
             );
 
             $this->template->parse(